Created PatientRecordRepository

+++ HospitalManagement.Core/Models/PatientRecord.cs
@@ -6,10 +6,10 @@ namespace HospitalManagement.Core.Models;
 
 public class PatientRecord
 {
-    public long Id { get; set; } // Primary Key
+    public int Id { get; set; } // Primary Key
 
     [Required (ErrorMessage = "PatientId is required")]
-    public long PatientId { get; set; } // Foreign Key: To which patient does this record belong to?
+    public int PatientId { get; set; } // Foreign Key: To which patient does this record belong to?
     public Patient? Patient { get; set; } // Navigation property
 
     public List<RecordEntry> Entries { get; set; } = new(); // Initially empty
